requirements.txt
pandas
numpy
joblib
scikit-learn==1.1.3
Flask
tensorflow-cpu
gunicorn